+++ database/migrations/2018_03_29_153535_create_table_temporadas_altas.php
@@ -20,7 +20,7 @@ class CreateTableTemporadasAltas extends Migration
             $table->integer('embarcacion_id')->unsigned();
             $table->date('desde');
             $table->date('hasta');
-            $table->decimal('cant_dias', 8, 2);
+            $table->integer('cant_dias');
             $table->decimal('gross', 8, 2);
             $table->decimal('neto', 8, 2);
             $table->decimal('comision_glc', 8, 2);
